# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API cho Superapp dành cho Tài xế công nghệ.\nBao gồm: Auth, GPS Tracking, SOS, Chat, Map Sinh tồn, Mood, Earning, Reminder, Emergency Contact.",
        "title": "Grab Driver Superapp API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/chat/conversations": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lấy danh sách conversation mà tài xế tham gia, mới nhất trước",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Danh sách conversation",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Số kết quả (mặc định 30)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset phân trang",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_modules_chat.Conversation"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Tạo chat 1-1 (direct) hoặc nhóm (group). Direct giữa 2 người nếu đã tồn tại sẽ trả lại cái cũ.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Tạo conversation",
                "parameters": [
                    {
                        "description": "Thông tin conversation",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_chat.CreateConversationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_chat.Conversation"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/chat/conversations/{convId}/messages": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lấy tin nhắn trong conversation, mới nhất trước",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Lịch sử tin nhắn",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Conversation ID",
                        "name": "convId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Số tin nhắn (mặc định 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset phân trang",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_modules_chat.Message"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gửi tin nhắn vào conversation qua REST API (backup cho WebSocket)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Gửi tin nhắn (REST)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Conversation ID (ObjectID hex)",
                        "name": "convId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Nội dung tin nhắn",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_chat.SendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_chat.Message"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/chat/ws": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upgrade lên WebSocket cho real-time chat. Gửi {\"type\":\"send\",\"conversation_id\":\"...\",\"content\":\"hello\"} để chat, {\"type\":\"typing\",\"conversation_id\":\"...\"} để báo đang gõ.",
                "tags": [
                    "Chat"
                ],
                "summary": "WebSocket Chat",
                "responses": {}
            }
        },
...
        "/earning/summary/daily": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Tổng thu, tổng chi, lãi ròng trong một ngày",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Earning"
                ],
                "summary": "Tổng kết ngày",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ngày YYYY-MM-DD (mặc định hôm nay)",
                        "name": "date",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_earning.DailySummary"
                        }
                    }
                }
            }
        },
        "/earning/summary/monthly": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Tổng thu, tổng chi, lãi ròng, trung bình lãi/ngày trong tháng",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Earning"
                ],
                "summary": "Tổng kết tháng",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tháng YYYY-MM (mặc định tháng này)",
                        "name": "month",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_earning.MonthlySummary"
                        }
                    }
                }
            }
        },
        "/earning/trips": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Lấy danh sách cuốc xe, lọc theo ngày nếu cần",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Earning"
                ],
                "summary": "Danh sách cuốc xe",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Lọc theo ngày YYYY-MM-DD",
                        "name": "date",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Số kết quả (mặc định 50)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_modules_earning.Trip"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Tài xế nhập cuốc xe mới với tổng thu, phí nền tảng, xăng, chi phí khác. Hệ thống tự tính lãi ròng.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Earning"
                ],
                "summary": "Ghi nhận cuốc xe",
                "parameters": [
                    {
                        "description": "Thông tin cuốc xe",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_earning.AddTripRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_earning.Trip"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_modules_chat.Conversation": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "description": "tên nhóm, rỗng nếu direct",
                    "type": "string"
                },
                "participants": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "type": {
                    "$ref": "#/definitions/internal_modules_chat.ConversationType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "internal_modules_chat.ConversationType": {
            "type": "string",
            "enum": [
                "direct",
                "group"
            ],
            "x-enum-comments": {
                "ConvDirect": "Chat 1-1",
                "ConvGroup": "Chat nhóm"
            },
            "x-enum-descriptions": [
                "Chat 1-1",
                "Chat nhóm"
            ],
            "x-enum-varnames": [
                "ConvDirect",
                "ConvGroup"
            ]
        },
        "internal_modules_chat.CreateConversationRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "participants": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "type": {
                    "$ref": "#/definitions/internal_modules_chat.ConversationType"
                }
            }
        },
        "internal_modules_chat.Message": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "content_type": {
                    "description": "\"text\", \"image\", \"voice\"",
                    "type": "string"
                },
                "conversation_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "sender_id": {
                    "type": "integer"
                }
            }
        },
        "internal_modules_chat.SendMessageRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "content_type": {
                    "description": "\"text\" nếu rỗng",
                    "type": "string"
                }
            }
        },
        "internal_modules_earning.AddTripRequest": {
            "type": "object",
            "properties": {
                "fuel_cost": {
                    "type": "number"
                },
                "gross_amount": {
                    "type": "number"
                },
                "note": {
                    "type": "string"
                },
                "other_cost": {
                    "type": "number"
                },
                "platform_fee": {
                    "type": "number"
                },
                "trip_date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                }
            }
        },
        "internal_modules_earning.DailySummary": {
            "type": "object",
            "properties": {
                "date": {
                    "description": "YYYY-MM-DD",
                    "type": "string"
                },
                "total_fees": {
                    "type": "number"
                },
                "total_fuel": {
                    "type": "number"
                },
                "total_gross": {
                    "type": "number"
                },
                "total_net": {
                    "type": "number"
                },
                "total_other": {
                    "type": "number"
                },
                "trip_count": {
                    "type": "integer"
                }
            }
        },
        "internal_modules_earning.MonthlySummary": {
            "type": "object",
            "properties": {
                "daily_avg_net": {
                    "description": "trung bình lãi ròng/ngày",
                    "type": "number"
                },
                "month": {
                    "description": "YYYY-MM",
                    "type": "string"
                },
                "total_fees": {
                    "type": "number"
                },
                "total_fuel": {
                    "type": "number"
                },
                "total_gross": {
                    "type": "number"
                },
                "total_net": {
                    "type": "number"
                },
                "total_other": {
                    "type": "number"
                },
                "trip_count": {
                    "type": "integer"
                }
            }
        },
        "internal_modules_earning.Trip": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "driver_id": {
                    "type": "integer"
                },
                "fuel_cost": {
                    "description": "xăng",
                    "type": "number"
                },
                "gross_amount": {
                    "description": "tổng thu",
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "net_amount": {
                    "description": "lãi ròng = gross - fees - costs",
                    "type": "number"
                },
                "note": {
                    "type": "string"
                },
                "other_cost": {
                    "description": "phí khác (gửi xe, phạt...)",
                    "type": "number"
                },
                "platform_fee": {
                    "description": "phí nền tảng (%)",
                    "type": "number"
                },
                "trip_date": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Nhập JWT token theo format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}